rename: lib 내 utils.ts를 cn.ts로 리네임
