+++ src/components/ui/Pagination/Pagination.styles.ts
@@ -1,4 +1,4 @@
-import { cn } from '@/lib/utils'
+import { cn } from '@/lib/cn'
 
 export function pageButton({
   active = false,
