+++ src/components/ui/Pagination/Pagination.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useMemo } from 'react'
 import { ChevronLeft, ChevronRight } from 'lucide-react'
-import { cn } from '@/lib/utils'
+import { cn } from '@/lib/cn'
 import type { PaginationProps } from './Pagination.types'
 import {
   clamp,
